See description (lots of changes)
added some better graphics for purify radial and abolish radial.  Fixed an error where killing someone while trip mined does not kill you.  Fixed an error where move again buffs don't show up properly on turn 0 (which technically won't happen in a real game, but I still would rather have that case accounted for).  Fixed an error that says that you wasted a purify or abolish (the message appeared if the first attempted square was already empty of buffs/debuffs instead of waiting to check all of them; this was due to a bad indentation).  Modified sleep times to be lower, making the game more responsive (this is possible due to the output messages from v0.9.10 allowing you to see messages that flashed by too quickly in the information box).  Temporarily disabled the title menu to save a click (and like 5 seconds per debug). Immediate to-do for next commit:  make the item menu neater.